treat ttl as 31 bits, per rfc 2181
git-svn-id: http://delta.affinix.com/svn/trunk/iris@241 ac7a513c-5e3a-0410-af42-968c3ac6e7dc

+++ src/jdns/jdns_util.c
@@ -1435,7 +1435,7 @@ jdns_rr_t *jdns_rr_from_resource(const jdns_packet_resource_t *pr, const jdns_pa
 	{
 		rr->qclass = pr->qclass;
 		rr->owner = _ustrdup(pr->qname->data);
-		rr->ttl = pr->ttl;
+		rr->ttl = (int)pr->ttl; // pr->ttl is 31-bits, cast is safe
 		rr->rdlength = pr->rdlength;
 		rr->rdata = jdns_copy_array(pr->rdata, pr->rdlength);
 	}
